Record post-b8f8bf4 natural dtor probe and fix lbl_806D2280 type
Reflect b8f8bf4 (canonical __dt__11ScopedTimerFv promoted to C++ shim) in
TODO.md and docs/notes/cpp-scoped-timer-pattern.md, and correct the
lbl_806D2280 extern decl from unsigned int to float (used by the natural
one-expression dtor probe).

Next blocker for natural dtor: GPR allocation
(target mulhwu r0,r6,r4 / divwu r0,r4,r0 vs current r4/r5 family).

+++ src/game/FlowDispatcher_ScopedTimer.c
@@ -17,7 +17,7 @@ extern "C" void SeqMenuScene_Init();
 /* --- extern decls: sda21-referenced data --- */
 extern "C" unsigned int lbl_806D0F80;
 extern "C" unsigned int lbl_806D2268;
-extern "C" unsigned int lbl_806D2280;
+extern "C" float lbl_806D2280;
 extern "C" unsigned int lbl_806D2288;
 
 /* --- extern decls: large-data refs (@ha/@l pairs) --- */
